restructured tella into a flat structure for the base module
