+++ examples/rl_minimal_agent.py
@@ -1,6 +1,6 @@
 import typing
-from tella.curriculum.rl_task_variant import Transition
-from tella.agents.continual_rl_agent import ContinualRLAgent, Observation, Action
+from tella.curriculum import Transition
+from tella.agents import ContinualRLAgent, Observation, Action
 
 
 class MinimalRandomAgent(ContinualRLAgent):
