--- tella/curriculum/rl_task_variant.py
@@ -1,209 +1,0 @@
-"""
-Copyright © 2021 The Johns Hopkins University Applied Physics Laboratory LLC
-
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the “Software”), to
-deal in the Software without restriction, including without limitation the
-rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
-sell copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
-IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-"""
-
-import typing
-import gym
-from ..env import L2LoggerEnv
-from .task_variant import AbstractTaskVariant
-from ..validation import validate_params
-
-Observation = typing.TypeVar("Observation")
-Action = typing.TypeVar("Action")
-Reward = float
-Done = bool
-NextObservation = Observation
-
-Transition = typing.Tuple[Observation, Action, Reward, Done, NextObservation]
-"""
-A tuple with data containing data from a single step in an MDP.
-The last item of the tuple is the observation resulting from applying the action
-to the observation (i.e. Next observation).
-"""
-
-ActionFn = typing.Callable[
-    [typing.List[typing.Optional[Observation]]], typing.List[typing.Optional[Action]]
-]
-"""
-A function that takes a list of Observations and returns a list of Actions, one
-for each observation.
-"""
-
-AbstractRLTaskVariant = AbstractTaskVariant[
-    ActionFn, typing.Iterable[Transition], gym.Env
-]
-"""
-An AbstractRLTaskVariant is an TaskVariant that takes an ActionFn as input
-and produces an Iterable[Transition]. It also  returns a :class:`gym.Env` as the
-Information.
-"""
-
-
-class EpisodicTaskVariant(AbstractRLTaskVariant):
-    """
-    Represents a TaskVariant that consists of a set number of episodes in a
-    :class:`gym.Env`.
-
-    This is a concrete subclass of the :class:`AbstractRLTaskVariant`,
-    that takes an :type:`ActionFn` and returns an iterable of :type:`Transition`.
-    """
-
-    def __init__(
-        self,
-        task_cls: typing.Type[gym.Env],
-        *,
-        num_episodes: int,
-        num_envs: typing.Optional[int] = None,
-        params: typing.Optional[typing.Dict] = None,
-        task_label: typing.Optional[str] = None,
-        variant_label: typing.Optional[str] = None,
-    ) -> None:
-        if num_envs is None:
-            num_envs = 1
-        if params is None:
-            params = {}
-        if task_label is None:
-            task_label = task_cls.__name__
-        if variant_label is None:
-            variant_label = "Default"
-        assert num_envs > 0
-
-        self._task_cls = task_cls
-        self._params = params
-        self._num_episodes = num_episodes
-        self._num_envs = num_envs
-        self._env = None
-        self._task_label = task_label
-        self._variant_label = variant_label
-        self.data_logger = None
-        self.logger_info = None
-
-    def total_episodes(self):
-        return self._num_episodes
-
-    @property
-    def task_label(self) -> str:
-        return self._task_label
-
-    @property
-    def variant_label(self) -> str:
-        return self._variant_label
-
-    def validate(self) -> None:
-        return validate_params(self._task_cls, list(self._params.keys()))
-
-    def _make_env(self) -> gym.Env:
-        """
-        Initializes the gym environment object and wraps in the L2MEnv to log rewards
-        """
-        if self.data_logger is not None:
-            return L2LoggerEnv(
-                self._task_cls(**self._params), self.data_logger, self.logger_info
-            )
-        else:
-            # FIXME: remove this after #31. this is to support getting spaces without setting l2logger info
-            return self._task_cls(**self._params)
-
-    def set_logger_info(
-        self, data_logger, block_num: int, is_learning_allowed: bool, exp_num: int
-    ):
-        self.data_logger = data_logger
-        self.logger_info = {
-            "block_num": block_num,
-            "block_type": "train" if is_learning_allowed else "test",
-            "task_params": self._params,
-            "task_name": self.task_label + "_" + self.variant_label,
-            "worker_id": "worker-default",
-            "exp_num": exp_num,
-        }
-
-    def info(self) -> gym.Env:
-        if self._env is None:
-            vector_env_cls = gym.vector.AsyncVectorEnv
-            if self._num_envs == 1:
-                vector_env_cls = gym.vector.SyncVectorEnv
-            self._env = vector_env_cls([self._make_env for _ in range(self._num_envs)])
-        return self._env
-
-    def generate(self, action_fn: ActionFn) -> typing.Iterable[Transition]:
-        env = self.info()
-
-        num_episodes_finished = 0
-
-        # data to keep track of which observations to mask out (set to None)
-        episode_ids = list(range(self._num_envs))
-        next_episode_id = episode_ids[-1] + 1
-
-        observations = env.reset()
-        while num_episodes_finished < self._num_episodes:
-            # mask out any environments that have episode id above max episodes
-            mask = [ep_id >= self._num_episodes for ep_id in episode_ids]
-
-            # replace masked environment observations with None
-            masked_observations = _where(mask, None, observations)
-
-            # query for the actions
-            actions = action_fn(masked_observations)
-
-            # replace masked environment actions with random action
-            unmasked_actions = _where(mask, env.single_action_space.sample(), actions)
-
-            # step in the VectorEnv
-            next_observations, rewards, dones, infos = env.step(unmasked_actions)
-
-            # yield all the non masked transitions
-            for i in range(self._num_envs):
-                if not mask[i]:
-                    # FIXME: if done[i] == True, then we need to return info[i]["terminal_observation"]
-                    yield (
-                        observations[i],
-                        actions[i],
-                        rewards[i],
-                        dones[i],
-                        next_observations[i],
-                    )
-
-                # increment episode ids if episode ended
-                if dones[i]:
-                    episode_ids[i] = next_episode_id
-                    next_episode_id += 1
-
-            observations = next_observations
-            num_episodes_finished += sum(dones)
-        self._env.close()
-        self._env = None
-
-
-def _where(
-    condition: typing.List[bool], replace_value: typing.Any, original_list: typing.List
-) -> typing.List:
-    """
-    Replaces elements in `original_list[i]` with `replace_value` where the `condition[i]`
-    is True.
-
-    :param condition: List of booleans indicating where to put replace_value
-    :param replace_value: The value to insert into the list
-    :param original_list: The list of values to modify
-    :return: A new list with replace_value inserted where condition elements are True
-    """
-    return [
-        replace_value if condition[i] else original_list[i]
-        for i in range(len(condition))
-    ]
