+++ tella/env.py
@@ -20,7 +20,9 @@ IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
 
 import typing
+
 import gym
+
 from l2logger import l2logger
