+++ tella/experiment.py
@@ -19,13 +19,14 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR
 IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
 
-import gym
 import logging
 import typing
-from .curriculum import AbstractCurriculum
-from .agents.continual_rl_agent import ContinualRLAgent, AbstractRLTaskVariant
-from .validation import validate_curriculum
-from .run import run
+
+import gym
+from l2logger import l2logger
+
+from .agents import ContinualRLAgent, ContinualLearningAgent, AbstractRLTaskVariant
+from .curriculum import AbstractCurriculum, AbstractTaskVariant, validate_curriculum
 
 
 logger = logging.getLogger(__name__)
@@ -146,3 +147,48 @@ def _spaces(
                 env.close()
                 del env
                 return observation_space, action_space
+
+
+def run(
+    agent: ContinualLearningAgent[AbstractTaskVariant],
+    curriculum: AbstractCurriculum[AbstractTaskVariant],
+):
+    """
+    Run an agent through an entire curriculum. This assumes that the agent
+    and the curriculum are both generic over the same type.
+
+    I.e. the curriculum will be generating task variants of type T, and the agent
+    will be consuming them via it's :meth:`ContinualLearningAgent.consume_task_variant`.
+    """
+    scenario_dir = curriculum.__class__.__name__
+    scenario_info = {
+        "author": "JHU APL",
+        "complexity": "1-low",
+        "difficulty": "2-medium",
+        "scenario_type": "custom",
+    }
+    logger_info = {"metrics_columns": ["reward"], "log_format_version": "1.0"}
+    # TODO change logs
+    data_logger = l2logger.DataLogger("logs", scenario_dir, logger_info, scenario_info)
+    total_episodes = 0
+    for i_block, block in enumerate(curriculum.learn_blocks_and_eval_blocks()):
+        is_learning_allowed = agent.is_learning_allowed = block.is_learning_allowed()
+        agent.block_start(is_learning_allowed)
+        for task_block in block.task_blocks():
+            agent.task_start(None)
+            for task_variant in task_block.task_variants():
+                # NOTE: assuming taskvariant has params
+                task_variant.set_logger_info(
+                    data_logger, i_block, is_learning_allowed, total_episodes
+                )
+                # FIXME: how to provide task & variant info?
+                agent.task_variant_start(None, None)
+                if is_learning_allowed:
+                    metrics = agent.learn_task_variant(task_variant)
+                else:
+                    metrics = agent.eval_task_variant(task_variant)
+                logger.info(f"TaskVariant produced metrics: {metrics}")
+                agent.task_variant_end(None, None)
+                total_episodes += task_variant.total_episodes()
+            agent.task_end(None)
+        agent.block_end(block.is_learning_allowed())
