+++ tests/simple_curriculum.py
@@ -1,8 +1,8 @@
 import typing
 import gym
 from tella.curriculum import AbstractCurriculum, AbstractLearnBlock, AbstractEvalBlock
-from tella.curriculum.rl_task_variant import AbstractRLTaskVariant, EpisodicTaskVariant
-from tella.curriculum.builders import simple_learn_block, simple_eval_block
+from tella.curriculum import AbstractRLTaskVariant, EpisodicTaskVariant
+from tella.curriculum import simple_learn_block, simple_eval_block
 
 
 class SimpleRLCurriculum(AbstractCurriculum[AbstractRLTaskVariant]):
