+++ tests/test_curriculum_validation.py
@@ -2,9 +2,9 @@ import pytest
 import typing
 import gym
 from tella.curriculum import AbstractCurriculum, AbstractLearnBlock, AbstractEvalBlock
-from tella.curriculum.rl_task_variant import AbstractRLTaskVariant, EpisodicTaskVariant
-from tella.curriculum.builders import simple_learn_block, simple_eval_block
-from tella.validation import validate_curriculum
+from tella.curriculum import AbstractRLTaskVariant, EpisodicTaskVariant
+from tella.curriculum import simple_learn_block, simple_eval_block
+from tella.curriculum import validate_curriculum
 
 
 class TestCurriculum(AbstractCurriculum[AbstractRLTaskVariant]):
